rework installation and update system

+++ src/share/sh/grub-installer_phase_1.sh
@@ -22,6 +22,14 @@ fi
 sharedir="$(dirname "$(dirname "$(realpath "$0")")")"
 clonetargetdir="$(realpath "$1")"
 
+
+if [ ! -f "$sharedir/sh/grub-installer_phase_2.sh" ]; then
+  echo "In target sys grub-installer_phase_2.sh isn't available"
+  echo "run install-installer first"
+  echo "Abort!"
+  exit 1
+fi
+
 if [ -e "${clonetargetdir}"/boot/grub/device.map ];then
   tempdev="$(sed -e "s/\((hd0)\)/# \1/" "${clonetargetdir}"/boot/grub/device.map)"
   echo "$tempdev" > "${clonetargetdir}"/boot/grub/device.map
@@ -34,24 +42,22 @@ tempprobegrub="$("$sharedir"/sh/devicefinder.sh dev "${clonetargetdir}" | sed -e
 echo "(hd0) ${tempprobegrub} #--specialclone-me--" >> "${clonetargetdir}"/boot/grub/device.map
 echo "finished"
 
+
+
+mount -o bind /proc "${clonedestdir}"/proc
+mount -o bind /sys "${clonedestdir}"/sys
+mount -o bind /dev "${clonedestdir}"/dev
+
 # display can be opened with tmp and run
-if [ $# -gt 1 ];then
-  chroot "${clonetargetdir}" "$sharedir"/sh/grub-installer_phase_2.sh
-else
-  chroot "${clonetargetdir}" "$sharedir"/sh/grub-installer_phase_2 "$2" "$3"
-fi
+mount -o bind /tmp "${clonedestdir}"/tmp
+mount -o bind /run "${clonedestdir}"/run
+shift # currently just one arg which must vanish
+chroot "${clonetargetdir}" "$sharedir"/sh/grub-installer_phase_2.sh "$@"
 echo "back from chroot"
+umount "${clonedestdir}"/{tmp,run,proc,sys,dev}
+echo "mounts cleaned up"
+
 tempsed=$(sed -e "/#--specialclone-me--/d" "${clonetargetdir}"/boot/grub/device.map)
 echo "$tempsed" > "${clonetargetdir}"/boot/grub/device.map
 #sed -i -e "s/# (hd0)/(hd0)/" "${clonetargetdir}"/boot/grub/device.map
 echo "device.map cleaned"
-if [[ $# = 1 ]];then
-  echo "if you want to use device.map please type \"yes\" and edit it now"
-  read shall_devicemap
-  if [ "$shall_devicemap" = "yes" ]; then
-    if ! ${EDITOR} "${clonetargetdir}"/boot/grub/device.map; then
-      echo "Fall back to vi"
-      vi "${clonetargetdir}"/boot/grub/device.map
-    fi
-  fi
-fi
